+++ docs/snippets/batch/accessProcessedMessages.ts
@@ -1,4 +1,4 @@
-import { BatchProcessor, EventType } from '@aws-lambda-powertools/batch';
+import { BatchProcessorSync, EventType } from '@aws-lambda-powertools/batch';
 import { Logger } from '@aws-lambda-powertools/logger';
 import type {
   SQSEvent,
@@ -7,7 +7,7 @@ import type {
   SQSBatchResponse,
 } from 'aws-lambda';
 
-const processor = new BatchProcessor(EventType.SQS);
+const processor = new BatchProcessorSync(EventType.SQS);
 const logger = new Logger();
 
 const recordHandler = (record: SQSRecord): void => {
@@ -25,7 +25,7 @@ export const handler = async (
   const batch = event.Records; // (1)!
 
   processor.register(batch, recordHandler, { context }); // (2)!
-  const processedMessages = processor.process();
+  const processedMessages = processor.processSync();
 
   for (const message of processedMessages) {
     const status: 'success' | 'fail' = message[0];
