+++ docs/snippets/batch/advancedTracingRecordHandler.ts
@@ -1,7 +1,7 @@
 import {
-  BatchProcessor,
+  BatchProcessorSync,
   EventType,
-  processPartialResponse,
+  processPartialResponseSync,
 } from '@aws-lambda-powertools/batch';
 import { Tracer, captureLambdaHandler } from '@aws-lambda-powertools/tracer';
 import middy from '@middy/core';
@@ -12,7 +12,7 @@ import type {
   SQSBatchResponse,
 } from 'aws-lambda';
 
-const processor = new BatchProcessor(EventType.SQS);
+const processor = new BatchProcessorSync(EventType.SQS);
 const tracer = new Tracer({ serviceName: 'serverlessAirline' });
 
 const recordHandler = (record: SQSRecord): void => {
@@ -36,7 +36,7 @@ const recordHandler = (record: SQSRecord): void => {
 
 export const handler = middy(
   async (event: SQSEvent, context: Context): Promise<SQSBatchResponse> => {
-    return processPartialResponse(event, recordHandler, processor, {
+    return processPartialResponseSync(event, recordHandler, processor, {
       context,
     });
   }
