+++ docs/snippets/batch/customPartialProcessor.ts
@@ -7,7 +7,7 @@ import { marshall } from '@aws-sdk/util-dynamodb';
 import {
   EventType,
   BasePartialBatchProcessor,
-  processPartialResponse,
+  processPartialResponseSync,
 } from '@aws-lambda-powertools/batch';
 import type {
   SuccessResponse,
@@ -25,12 +25,6 @@ class MyPartialProcessor extends BasePartialBatchProcessor {
   public constructor(tableName: string) {
     super(EventType.SQS);
     this.#tableName = tableName;
-  }
-
-  public async asyncProcessRecord(
-    _record: BaseRecord
-  ): Promise<SuccessResponse | FailureResponse> {
-    throw new Error('Not implemented');
   }
 
   /**
@@ -64,13 +58,21 @@ class MyPartialProcessor extends BasePartialBatchProcessor {
     this.successMessages = [];
   }
 
+  public async processRecord(
+    _record: BaseRecord
+  ): Promise<SuccessResponse | FailureResponse> {
+    throw new Error('Not implemented');
+  }
+
   /**
    * It handles how your record is processed.
    *
    * Here we are keeping the status of each run, `this.handler` is
    * the function that is passed when calling `processor.register()`.
    */
-  public processRecord(record: BaseRecord): SuccessResponse | FailureResponse {
+  public processRecordSync(
+    record: BaseRecord
+  ): SuccessResponse | FailureResponse {
     try {
       const result = this.handler(record);
 
@@ -91,7 +93,7 @@ export const handler = async (
   event: SQSEvent,
   context: Context
 ): Promise<SQSBatchResponse> => {
-  return processPartialResponse(event, recordHandler, processor, {
+  return processPartialResponseSync(event, recordHandler, processor, {
     context,
   });
 };
