+++ docs/snippets/batch/gettingStartedDynamoDBStreams.ts
@@ -1,7 +1,7 @@
 import {
-  BatchProcessor,
+  BatchProcessorSync,
   EventType,
-  processPartialResponse,
+  processPartialResponseSync,
 } from '@aws-lambda-powertools/batch';
 import { Logger } from '@aws-lambda-powertools/logger';
 import type {
@@ -11,7 +11,7 @@ import type {
   DynamoDBBatchResponse,
 } from 'aws-lambda';
 
-const processor = new BatchProcessor(EventType.DynamoDBStreams); // (1)!
+const processor = new BatchProcessorSync(EventType.DynamoDBStreams); // (1)!
 const logger = new Logger();
 
 const recordHandler = (record: DynamoDBRecord): void => {
@@ -29,7 +29,7 @@ export const handler = async (
   event: DynamoDBStreamEvent,
   context: Context
 ): Promise<DynamoDBBatchResponse> => {
-  return processPartialResponse(event, recordHandler, processor, {
+  return processPartialResponseSync(event, recordHandler, processor, {
     context,
   });
 };
