+++ packages/batch/src/BatchProcessor.ts
@@ -6,26 +6,30 @@ import type { BaseRecord, FailureResponse, SuccessResponse } from './types';
  * Process native partial responses from SQS, Kinesis Data Streams, and DynamoDB
  */
 class BatchProcessor extends BasePartialBatchProcessor {
-  public async asyncProcessRecord(
-    _record: BaseRecord
+  public async processRecord(
+    record: BaseRecord
   ): Promise<SuccessResponse | FailureResponse> {
-    throw new BatchProcessingError('Not implemented. Use process() instead.');
-  }
-
-  /**
-   * Process a record with instance's handler
-   * @param record Batch record to be processed
-   * @returns response of success or failure
-   */
-  public processRecord(record: BaseRecord): SuccessResponse | FailureResponse {
     try {
       const data = this.toBatchType(record, this.eventType);
-      const result = this.handler(data, this.options?.context);
+      const result = await this.handler(data, this.options?.context);
 
       return this.successHandler(record, result);
     } catch (error) {
       return this.failureHandler(record, error as Error);
     }
+  }
+
+  /**
+   * Process a record with instance's handler
+   * @param _record Batch record to be processed
+   * @returns response of success or failure
+   */
+  public processRecordSync(
+    _record: BaseRecord
+  ): SuccessResponse | FailureResponse {
+    throw new BatchProcessingError(
+      'Not implemented. Use asyncProcess() instead.'
+    );
   }
 }
