+++ packages/batch/src/index.ts
@@ -3,8 +3,8 @@ export * from './errors';
 export * from './types';
 export * from './BasePartialProcessor';
 export * from './BasePartialBatchProcessor';
+export * from './BatchProcessorSync';
 export * from './BatchProcessor';
-export * from './AsyncBatchProcessor';
+export * from './processPartialResponseSync';
 export * from './processPartialResponse';
-export * from './asyncProcessPartialResponse';
 export * from './SqsFifoPartialProcessor';
