+++ packages/batch/tests/unit/processPartialResponse.test.ts
@@ -1,7 +1,7 @@
 /**
- * Test processPartialResponse function
+ * Test asyncProcessPartialResponse function
  *
- * @group unit/batch/function/processpartialresponse
+ * @group unit/batch/function/asyncProcesspartialresponse
  */
 import type {
   Context,
@@ -11,7 +11,8 @@ import type {
 } from 'aws-lambda';
 import { helloworldContext as dummyContext } from '@aws-lambda-powertools/commons/lib/samples/resources/contexts';
 import { Custom as dummyEvent } from '@aws-lambda-powertools/commons/lib/samples/resources/events';
-import { BatchProcessor, processPartialResponse } from '../../src';
+import { BatchProcessor } from '../../src/BatchProcessor';
+import { processPartialResponse } from '../../src/processPartialResponse';
 import { EventType } from '../../src/constants';
 import type {
   BatchProcessingOptions,
@@ -23,10 +24,10 @@ import {
   sqsRecordFactory,
 } from '../helpers/factories';
 import {
-  dynamodbRecordHandler,
-  handlerWithContext,
-  kinesisRecordHandler,
-  sqsRecordHandler,
+  asyncDynamodbRecordHandler,
+  asyncHandlerWithContext,
+  asyncKinesisRecordHandler,
+  asyncSqsRecordHandler,
 } from '../helpers/handlers';
 
 describe('Function: processPartialResponse()', () => {
@@ -45,7 +46,7 @@ describe('Function: processPartialResponse()', () => {
   });
 
   describe('Process partial response function call tests', () => {
-    test('Process partial response function call with synchronous handler', () => {
+    test('Process partial response function call with asynchronous handler', async () => {
       // Prepare
       const records = [
         sqsRecordFactory('success'),
@@ -55,13 +56,17 @@ describe('Function: processPartialResponse()', () => {
       const processor = new BatchProcessor(EventType.SQS);
 
       // Act
-      const ret = processPartialResponse(batch, sqsRecordHandler, processor);
+      const ret = await processPartialResponse(
+        batch,
+        asyncSqsRecordHandler,
+        processor
+      );
 
       // Assess
       expect(ret).toStrictEqual({ batchItemFailures: [] });
     });
 
-    test('Process partial response function call with context provided', () => {
+    test('Process partial response function call with context provided', async () => {
       // Prepare
       const records = [
         sqsRecordFactory('success'),
@@ -71,9 +76,9 @@ describe('Function: processPartialResponse()', () => {
       const processor = new BatchProcessor(EventType.SQS);
 
       // Act
-      const ret = processPartialResponse(
+      const ret = await processPartialResponse(
         batch,
-        handlerWithContext,
+        asyncHandlerWithContext,
         processor,
         options
       );
@@ -84,7 +89,7 @@ describe('Function: processPartialResponse()', () => {
   });
 
   describe('Process partial response function call through handler', () => {
-    test('Process partial response through handler with SQS event', () => {
+    test('Process partial response through handler with SQS event', async () => {
       // Prepare
       const records = [
         sqsRecordFactory('success'),
@@ -93,21 +98,21 @@ describe('Function: processPartialResponse()', () => {
       const processor = new BatchProcessor(EventType.SQS);
       const event: SQSEvent = { Records: records };
 
-      const handler = (
+      const handler = async (
         event: SQSEvent,
         _context: Context
-      ): PartialItemFailureResponse => {
-        return processPartialResponse(event, sqsRecordHandler, processor);
+      ): Promise<PartialItemFailureResponse> => {
+        return processPartialResponse(event, asyncSqsRecordHandler, processor);
       };
 
       // Act
-      const result = handler(event, context);
+      const result = await handler(event, context);
 
       // Assess
       expect(result).toStrictEqual({ batchItemFailures: [] });
     });
 
-    test('Process partial response through handler with Kinesis event', () => {
+    test('Process partial response through handler with Kinesis event', async () => {
       // Prepare
       const records = [
         kinesisRecordFactory('success'),
@@ -116,21 +121,25 @@ describe('Function: processPartialResponse()', () => {
       const processor = new BatchProcessor(EventType.KinesisDataStreams);
       const event: KinesisStreamEvent = { Records: records };
 
-      const handler = (
+      const handler = async (
         event: KinesisStreamEvent,
         _context: Context
-      ): PartialItemFailureResponse => {
-        return processPartialResponse(event, kinesisRecordHandler, processor);
+      ): Promise<PartialItemFailureResponse> => {
+        return await processPartialResponse(
+          event,
+          asyncKinesisRecordHandler,
+          processor
+        );
       };
 
       // Act
-      const result = handler(event, context);
+      const result = await handler(event, context);
 
       // Assess
       expect(result).toStrictEqual({ batchItemFailures: [] });
     });
 
-    test('Process partial response through handler with DynamoDB event', () => {
+    test('Process partial response through handler with DynamoDB event', async () => {
       // Prepare
       const records = [
         dynamodbRecordFactory('success'),
@@ -139,41 +148,51 @@ describe('Function: processPartialResponse()', () => {
       const processor = new BatchProcessor(EventType.DynamoDBStreams);
       const event: DynamoDBStreamEvent = { Records: records };
 
-      const handler = (
+      const handler = async (
         event: DynamoDBStreamEvent,
         _context: Context
-      ): PartialItemFailureResponse => {
-        return processPartialResponse(event, dynamodbRecordHandler, processor);
+      ): Promise<PartialItemFailureResponse> => {
+        return await processPartialResponse(
+          event,
+          asyncDynamodbRecordHandler,
+          processor
+        );
       };
 
       // Act
-      const result = handler(event, context);
+      const result = await handler(event, context);
 
       // Assess
       expect(result).toStrictEqual({ batchItemFailures: [] });
     });
 
-    test('Process partial response through handler for SQS records with incorrect event type', () => {
+    test('Process partial response through handler for SQS records with incorrect event type', async () => {
       // Prepare
       const processor = new BatchProcessor(EventType.SQS);
       const event = dummyEvent;
 
-      const handler = (
+      const handler = async (
         event: SQSEvent,
         _context: Context
-      ): PartialItemFailureResponse => {
-        return processPartialResponse(event, sqsRecordHandler, processor);
+      ): Promise<PartialItemFailureResponse> => {
+        return await processPartialResponse(
+          event,
+          asyncSqsRecordHandler,
+          processor
+        );
       };
 
       // Act & Assess
-      expect(() => handler(event as unknown as SQSEvent, context)).toThrowError(
+      await expect(() =>
+        handler(event as unknown as SQSEvent, context)
+      ).rejects.toThrowError(
         `Unexpected batch type. Possible values are: ${Object.keys(
           EventType
         ).join(', ')}`
       );
     });
 
-    test('Process partial response through handler with context provided', () => {
+    test('Process partial response through handler with context provided', async () => {
       // Prepare
       const records = [
         sqsRecordFactory('success'),
@@ -182,22 +201,22 @@ describe('Function: processPartialResponse()', () => {
       const processor = new BatchProcessor(EventType.SQS);
       const event: SQSEvent = { Records: records };
 
-      const handler = (
+      const handler = async (
         event: SQSEvent,
         context: Context
-      ): PartialItemFailureResponse => {
+      ): Promise<PartialItemFailureResponse> => {
         const options: BatchProcessingOptions = { context: context };
 
-        return processPartialResponse(
+        return await processPartialResponse(
           event,
-          handlerWithContext,
+          asyncHandlerWithContext,
           processor,
           options
         );
       };
 
       // Act
-      const result = handler(event, context);
+      const result = await handler(event, context);
 
       // Assess
       expect(result).toStrictEqual({ batchItemFailures: [] });
